Add input validation for leave records

+++ sqlite_excel.py
@@ -41,7 +41,7 @@ def convert_input(upload_file):
     neft_incoming = pd.read_excel(upload_file, skiprows=4, usecols=range(1, 16))
     neft_incoming = neft_incoming[neft_incoming["File Splited"] != "Yes"]
     neft_incoming = neft_incoming[neft_incoming["Download Status"] != "Downloaded"]
-    df_db["AMOUNT_db"] = df_db["AMOUNT_db"].astype(int)
+    df_db["AMOUNT_db"] = df_db["AMOUNT_db"].astype(float)
     neft_incoming["Reference Date"] = pd.to_datetime(
         neft_incoming["Reference Date"], dayfirst=True
     )
